ref: sepearted slot_iterator from slot
