+++ src/models/mod.rs
@@ -3,5 +3,6 @@ pub mod input;
 pub mod output;
 pub mod repetition;
 pub mod slot;
+pub mod slot_iterator;
 pub mod task;
 pub mod task_budgets;
