+++ src/models/slot_iterator/iterator.rs
