+++ src/models/slot_iterator/mod.rs
@@ -1,0 +1,40 @@
+pub mod filter;
+pub mod iterator;
+pub mod utils;
+
+use super::{goal::TimeFilter, repetition::Repetition, slot::Slot};
+use chrono::NaiveDateTime;
+
+// derive Debug for TimeSlotsIterator
+#[derive(Debug)]
+/// An iterator that returns slots that conform to a repetition,
+/// with optional filters (after/before/Mondays/Weekdays/Weekends/Bank holidays...)
+/// e.g. iterate over all MONDAYS between 1st September 2022 and 30th September 2022.
+/// e.g. iterate over all DAYS between 1st September 2022 and 30th September 2022.
+/// e.g. iterate over all HOURS between 1st September 2022 and 30th September 2022.
+/// e.g. iterate over all 'every two hours' slots between times 10-21 for time period X-Y
+pub(crate) struct TimeSlotsIterator {
+    timeline: Vec<Slot>,
+    repetition: Option<Repetition>,
+    filters: Option<TimeFilter>,
+    current_start_position: NaiveDateTime,
+}
+
+impl TimeSlotsIterator {
+    pub fn new(
+        start: NaiveDateTime,
+        end: NaiveDateTime,
+        repetition: Option<Repetition>,
+        filters: Option<TimeFilter>,
+    ) -> TimeSlotsIterator {
+        let mut result = TimeSlotsIterator {
+            timeline: vec![Slot { start, end }],
+            repetition,
+            filters,
+            current_start_position: start, //override after applying filters
+        };
+        result.apply_filters();
+        result.current_start_position = result.timeline[0].start;
+        result
+    }
+}
